--- src/components/SelectTransacCategory/TransCategory.module.css
@@ -1,28 +1,0 @@
-.groupStyles {
-  margin-top: 40px;
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  margin-top: 40px;
-  margin-bottom: 40px;
-  width: 100%;
-  height: 34px;
-
-  border-bottom: 1px solid #e0e0e0;
-  text-decoration-color: #bdbdbd;
-}
-
-.groupBadgeStyles {
-  background: rgba(255, 255, 255, 0.7);
-  box-shadow: 0px 6px 15px rgba(0, 0, 0, 0.1);
-  backdrop-filter: blur(50px);
-  border-radius: 20px;
-  display: inline-block;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 18px;
-  line-height: 27px;
-  display: flex;
-  align-items: flex-start;
-  color: #000000;
-}
